requirements.txt
# TimesFM - Modèle de base
timesfm[torch]>=2.0.0

# Données et manipulation
pandas>=2.0.0
numpy>=1.24.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# APIs météo
requests>=2.31.0

# Utilitaires data science
scikit-learn>=1.3.0
scipy>=1.11.0

# Jupyter pour exploration
jupyter>=1.0.0
ipykernel>=6.25.0

# Configuration et logging
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Tests
pytest>=7.4.0
pytest-cov>=4.1.0